Basic skeleton of chain separation
The change use a new configuration option to determine which kind of
stratum agent server to create. The underlying structure is dummy and
actual implementations are to be done.
