+++ CMakeLists.txt
@@ -40,7 +40,9 @@ include_directories(src test ${GLOG_INCLUDE_DIRS} ${LIBEVENT_INCLUDE_DIR})
 set(THRID_LIBRARIES -lpthread ${GLOG_LIBRARIES} ${LIBEVENT_LIB})
 
 file(GLOB LIB_SOURCES src/*.cc src/*.c)
-add_library(btccomagent STATIC ${LIB_SOURCES})
+set(LIB_BITCOIN_SOURCES src/bitcoin/ServerBitcoin.cc src/bitcoin/ServerBitcoin.h)
+set(LIB_ETH_SOURCES src/eth/ServerEth.cc src/eth/ServerEth.h)
+add_library(btccomagent STATIC ${LIB_SOURCES} ${LIB_BITCOIN_SOURCES} ${LIB_ETH_SOURCES})
 
 file(GLOB_RECURSE TEST_SOURCES test/*.cc)
 add_executable(unittest ${TEST_SOURCES})
