+++ CMakeLists4EmbeddedSystem.txt
@@ -32,6 +32,8 @@ link_directories(${SDK_DIR}/lib)
 set(THRID_LIBRARIES -ldl -lpthread -levent)
 
 file(GLOB LIB_SOURCES src/*.cc src/*.c)
+set(LIB_BITCOIN_SOURCES src/bitcoin/ServerBitcoin.cc src/bitcoin/ServerBitcoin.h)
+set(LIB_ETH_SOURCES src/eth/ServerEth.cc src/eth/ServerEth.h)
 add_library(btccomagent STATIC ${LIB_SOURCES})
 
 file(GLOB_RECURSE TEST_SOURCES test/*.cc)
