+++ CMakeLists4Windows.txt
@@ -125,6 +125,8 @@ include_directories(src test ${GLOG_INCLUDE_DIRS} ${LIBEVENT_INCLUDE_DIR})
 set(THRID_LIBRARIES ${GLOG_LIBRARIES} ${LIBEVENT_LIB})
 
 file(GLOB LIB_SOURCES src/*.cc src/*.c)
+set(LIB_BITCOIN_SOURCES src/bitcoin/ServerBitcoin.cc src/bitcoin/ServerBitcoin.h)
+set(LIB_ETH_SOURCES src/eth/ServerEth.cc src/eth/ServerEth.h)
 add_library(btccomagent STATIC ${LIB_SOURCES})
 
 file(GLOB_RECURSE TEST_SOURCES test/*.cc)
