+++ src/Server.cc
@@ -92,15 +92,6 @@ bool tryReadLine(string &line, struct evbuffer *inBuf) {
   return true;
 }
 
-static
-string getWorkerName(const string &fullName) {
-  size_t pos = fullName.find(".");
-  if (pos == fullName.npos) {
-    return "";
-  }
-  return fullName.substr(pos + 1);  // not include '.'
-}
-
 
 //////////////////////////////// StratumError ////////////////////////////////
 const char * StratumError::toString(int err) {
@@ -188,10 +179,8 @@ void SessionIDManager::freeSessionId(const uint16_t sessionId) {
 
 ///////////////////////////////// StratumMessage //////////////////////////////
 StratumMessage::StratumMessage(const string &content):
-content_(content), isStringId_(false), r_(0), diff_(0) {
+content_(content), isStringId_(false), r_(0) {
   parse();
-}
-StratumMessage::~StratumMessage() {
 }
 
 string StratumMessage::getJsonStr(const jsmntok_t *t) const {
@@ -229,187 +218,11 @@ string StratumMessage::parseId() {
   return "";
 }
 
-void StratumMessage::_parseMiningSubmit() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // {"params": ["slush.miner1", "bf", "00000001", "504e86ed", "b2957c02"],
-    //  "id": 4, "method": "mining.submit"}
-    //
-    // [Worker Name, Job ID, ExtraNonce2(hex), nTime(hex), nonce(hex)]
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size >= 5) {
-      bool hasVersionMask = t_[i+1].size >= 6;
-
-      i++;  // ptr move to params
-      i++;  // ptr move to params[0]
-
-      // Job ID
-      share_.jobId_       = (uint8_t) strtoul(getJsonStr(&t_[i+1]).c_str(), NULL, 10);
-      // ExtraNonce2(hex)
-      share_.extraNonce2_ = (uint32_t)strtoul(getJsonStr(&t_[i+2]).c_str(), NULL, 16);
-      // nTime(hex)
-      share_.time_        = (uint32_t)strtoul(getJsonStr(&t_[i+3]).c_str(), NULL, 16);
-      // nonce(hex)
-      share_.nonce_       = (uint32_t)strtoul(getJsonStr(&t_[i+4]).c_str(), NULL, 16);
-
-      if (hasVersionMask) {
-        // versionMask(hex)
-        share_.versionMask_ = (uint32_t)strtoul(getJsonStr(&t_[i+5]).c_str(), NULL, 16);
-      }
-
-      // set the method_
-      method_ = "mining.submit";
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningNotify() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // 9 elements in parmas
-    // "params": [
-    // "",           // Job ID
-    // "",           // Hash of previous block
-    // "",           // coinbase1
-    // "",           // coinbase2
-    // [],           // list of merkle branches
-    // "00000002",   // block version, hex
-    // "1c2ac4af",   // bits, hex
-    // "504e86b9",   // time, hex
-    // false]        // is_clean
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size == 9) {
-      i++;  // ptr move to params
-      i++;  // ptr move to params[0]
-      sjob_.jobId_    = (uint8_t)strtoul(getJsonStr(&t_[i]).c_str(), NULL, 10);
-      sjob_.prevHash_ = getJsonStr(&t_[i+1]);
-
-      // list of merkle branches
-      i += 4;  // ptr move to params[4]: list of merkle branches
-      if (t_[i].type != JSMN_ARRAY)
-        return;  // should be array
-
-      i += t_[i].size + 1;  // move to params[5]
-
-      sjob_.version_  = (int32_t) strtoul(getJsonStr(&t_[i]).c_str(),   NULL, 16);
-      sjob_.time_     = (uint32_t)strtoul(getJsonStr(&t_[i+2]).c_str(), NULL, 16);
-
-      const string isClean = str2lower(getJsonStr(&t_[i+3]));
-      sjob_.isClean_  = (isClean == "true") ? true : false;
-
-      // set the method_
-      method_ = "mining.notify";
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningSetDifficulty() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // { "id": null, "method": "mining.set_difficulty", "params": [2]}
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size == 1) {
-      i++;  // ptr move to params
-      i++;  // ptr move to params[0]
-      const uint32_t diff = (uint32_t)strtoul(getJsonStr(&t_[i]).c_str(), NULL, 10);
-
-      if (diff > 0) {
-        diff_ = diff;
-        // set the method_
-        method_ = "mining.set_difficulty";
-      }
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningSetVersionMask() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // {"id":null,"method":"mining.set_version_mask","params":["1fffe000"]}
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size == 1) {
-      i++;  // ptr move to params
-      i++;  // ptr move to params[0]
-      const uint32_t versionMask = (uint32_t)strtoul(getJsonStr(&t_[i]).c_str(), NULL, 16);
-
-      if (versionMask > 0) {
-        versionMask_ = versionMask;
-        // set the method_
-        method_ = "mining.set_version_mask";
-      }
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningSubscribe() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // {"id": 1, "method": "mining.subscribe", "params": ["bfgminer/4.4.0-32-gac4e9b3", "01ad557d"]}
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY) {
-      // some miners will use empty "params"
-      if (t_[i+1].size >= 1) {
-        i++;  // ptr move to params
-        i++;  // ptr move to params[0]
-        minerAgent_ = getJsonStr(&t_[i]);
-      }
-
-      // set the method_
-      method_ = "mining.subscribe";
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningAuthorize() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // {"params": ["slush.miner1", "password"], "id": 2, "method": "mining.authorize"}
-    //
-    if (jsoneq(&t_[i], "params") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size >= 1) {
-      i++;  // ptr move to params
-      i++;  // ptr move to params[0]
-      workerName_ = getJsonStr(&t_[i]);
-
-      // set the method_
-      method_ = "mining.authorize";
-      break;
-    }
-  }
-}
-
-void StratumMessage::_parseMiningConfigure() {
-  for (int i = 1; i < r_; i++) {
-    //
-    // {"id":3,"method":"mining.configure","params":[
-    //    ["version-rolling"],
-    //    {"version-rolling.mask":"1fffe000","version-rolling.min-bit-count":2}
-    // ]}
-    //
-    if (jsoneq(&t_[i], "version-rolling.mask") == 0 && t_[i+1].type == JSMN_STRING) {
-      i++;  // ptr move to value
-      const uint32_t versionMask = (uint32_t)strtoul(getJsonStr(&t_[i]).c_str(), NULL, 16);
-
-      if (versionMask > 0) {
-        versionMask_ = versionMask;
-
-        // set the method_
-        method_ = "mining.configure";
-        break;
-      }
-    }
-  }
-}
-
 void StratumMessage::parse() {
   jsmn_parser p;
   jsmn_init(&p);
 
-  r_ = jsmn_parse(&p, content_.c_str(), content_.length(), t_, sizeof(t_)/sizeof(t_[0]));
+  r_ = jsmn_parse(&p, content_.c_str(), content_.length(), t_, sizeof(t_) / sizeof(t_[0]));
   if (r_ < 0) {
     LOG(ERROR) << "failed to parse JSON: " << r_ << std::endl;
     return;
@@ -420,46 +233,6 @@ void StratumMessage::parse() {
     return;
 
   parseId();
-
-  // find method name
-  const string method = findMethod();
-  if (method == "")
-    return;
-
-  if (method == "mining.submit") {
-    _parseMiningSubmit();
-    return;
-  }
-
-  if (method == "mining.notify") {
-    _parseMiningNotify();
-    return;
-  }
-
-  if (method == "mining.set_difficulty") {
-    _parseMiningSetDifficulty();
-    return;
-  }
-
-  if (method == "mining.set_version_mask") {
-    _parseMiningSetVersionMask();
-    return;
-  }
-
-  if (method == "mining.subscribe") {
-    _parseMiningSubscribe();
-    return;
-  }
-
-  if (method == "mining.authorize") {
-    _parseMiningAuthorize();
-    return;
-  }
-
-  if (method == "mining.configure") {
-    _parseMiningConfigure();
-    return;
-  }
 }
 
 string StratumMessage::getMethod() const {
@@ -488,107 +261,10 @@ bool StratumMessage::isValid() const {
   return (r_ < 1 || t_[0].type != JSMN_OBJECT) ? false : true;
 }
 
-bool StratumMessage::parseMiningSubmit(Share &share) const {
-  if (method_ != "mining.submit")
-    return false;
-  share = share_;
-  return true;
-}
-
-bool StratumMessage::parseMiningSubscribe(string &minerAgent) const {
-  if (method_ != "mining.subscribe")
-    return false;
-  minerAgent = minerAgent_;
-  return true;
-}
-
-bool StratumMessage::parseMiningAuthorize(string &workerName) const {
-  if (method_ != "mining.authorize")
-    return false;
-  workerName = workerName_;
-  return true;
-
-}
-bool StratumMessage::parseMiningNotify(StratumJob &sjob) const {
-  if (method_ != "mining.notify")
-    return false;
-  sjob = sjob_;
-  return true;
-}
-bool StratumMessage::parseMiningSetDifficulty(uint32_t *diff) const {
-  if (method_ != "mining.set_difficulty")
-    return false;
-  *diff = diff_;
-  return true;
-}
-
-bool StratumMessage::parseMiningSetVersionMask(uint32_t *versionMask) const {
-  if (method_ != "mining.set_version_mask")
-    return false;
-  *versionMask = versionMask_;
-  return true;
-}
-
-bool StratumMessage::parseMiningConfigure(uint32_t *versionMask) const {
-  if (method_ != "mining.configure")
-    return false;
-  *versionMask = versionMask_;
-  return true;
-}
-
-bool StratumMessage::getExtraNonce1AndExtraNonce2Size(uint32_t *nonce1,
-                                                      int32_t *n2size) const {
-  bool r = false;
-  for (int i = 1; i < r_; i++) {
-    //
-    // 1. Subscriptions details
-    // 2. Extranonce1 - Hex-encoded
-    // 3. Extranonce2_size
-    //
-    // {"id":1,"result":[[["mining.set_difficulty","01000002"],
-    //                    ["mining.notify","01000002"]],"01000002",8],"error":null}
-    //
-    if (jsoneq(&t_[i], "result") == 0 && t_[i+1].type == JSMN_ARRAY && t_[i+1].size == 3) {
-      i++;  // ptr move to "result"
-
-      //
-      // subscriptions details: 2-tuple
-      //
-      i++;  // ptr move to result[0]
-      i++;  // ptr move to result[0][0]
-      // ["mining.set_difficulty","01000002"]
-      if (!(t_[i].type == JSMN_ARRAY && t_[i].size == 2)) {
-        return false;
-      }
-      i += t_[i].size;
-
-      i++;  // ptr move to result[0][1]
-      // ["mining.notify","01000002"]
-      if (!(t_[i].type == JSMN_ARRAY && t_[i].size == 2)) {
-        return false;
-      }
-      i += t_[i].size;
-
-      // extranonce1, hex
-      i++;  // ptr move to result[1]
-      *nonce1 = (uint32_t)strtoul(getJsonStr(&t_[i]).c_str(), NULL, 16);
-
-      // Extranonce2_size
-      i++;  // ptr move to result[2]
-      *n2size = (int32_t)strtol(getJsonStr(&t_[i]).c_str(), NULL, 10);
-
-      r = true;
-      break;
-    }
-  }
-  return r;
-}
-
-
 ///////////////////////////////// UpStratumClient //////////////////////////////
 UpStratumClient::UpStratumClient(const int8_t idx, struct event_base *base,
                                  const string &userName, StratumServer *server)
-: state_(UP_INIT), idx_(idx), server_(server), poolDefaultDiff_(0), versionMask_(0)
+: state_(UP_INIT), idx_(idx), server_(server), poolDefaultDiff_(0)
 {
   bev_ = bufferevent_socket_new(base, -1, BEV_OPT_CLOSE_ON_FREE);
   assert(bev_ != NULL);
@@ -602,11 +278,7 @@ UpStratumClient::UpStratumClient(const int8_t idx, struct event_base *base,
   bufferevent_enable(bev_, EV_READ|EV_WRITE);
 
   extraNonce1_ = 0u;
-  extraNonce2_ = 0u;
   userName_ = userName;
-
-  latestJobId_[0] = latestJobId_[1] = latestJobId_[2] = 0;
-  latestJobGbtTime_[0] = latestJobGbtTime_[1] = latestJobGbtTime_[2] = 0;
 
   lastJobReceivedTime_ = 0u;
 
@@ -683,153 +355,16 @@ bool UpStratumClient::handleMessage() {
   return false;  // read mesasge failure
 }
 
-void UpStratumClient::handleExMessage_MiningSetDiff(const string *exMessage) {
-  //
-  // CMD_MINING_SET_DIFF
-  // | magic_number(1) | cmd(1) | len (2) | diff_2_exp(1) | count(2) | session_id (2) ... |
-  //
-  const uint8_t *p = (uint8_t *)exMessage->data();
-  const uint8_t diff_2exp = *(p + 4);
-  const uint64_t diff = (uint64_t)exp2(diff_2exp);
-
-  const uint16_t count   = *(uint16_t *)(p + 5);
-  uint16_t *sessionIdPtr =  (uint16_t *)(p + 7);
-
-  for (size_t i = 0; i < count; i++) {
-    uint16_t sessionId = *sessionIdPtr++;
-    server_->sendMiningDifficulty(this, sessionId, diff);
-  }
-
-  LOG(INFO) << "up[" << (int32_t)idx_ << "] CMD_MINING_SET_DIFF, diff: "
-  << diff << ", sessions count: " << count << std::endl;
-}
-
 void UpStratumClient::sendData(const char *data, size_t len) {
   // add data to a bufferevent’s output buffer
   bufferevent_write(bev_, data, len);
 //  DLOG(INFO) << "UpStratumClient send(" << len << "): " << data << std::endl;
 }
 
-void UpStratumClient::sendMiningNotify(const string &line) {
-  // send to all down sessions
-  server_->sendMiningNotifyToAll(idx_, latestMiningNotifyStr_);
-}
-
-void UpStratumClient::convertMiningNotifyStr(const string &line) {
-  const char *pch = splitNotify(line);
-  latestMiningNotifyStr_.clear();
-
-  latestMiningNotifyStr_.append(line.c_str(), pch - line.c_str());
-  latestMiningNotifyStr_.append(Strings::Format("%08x", extraNonce1_));
-  latestMiningNotifyStr_.append(pch);
-}
-
-void UpStratumClient::handleStratumMessage(const string &line) {
-  DLOG(INFO) << "UpStratumClient recv(" << line.size() << "): " << line << std::endl;
-
-  StratumMessage smsg(line);
-  if (!smsg.isValid()) {
-    LOG(ERROR) << "decode line fail, not a json string" << std::endl;
-    return;
-  }
-
-  const string method = smsg.getMethod();
-  StratumJob sjob;
-  uint32_t difficulty = 0u;
-  uint32_t versionMask = 0u;
-
-  if (state_ == UP_AUTHENTICATED) {
-    if (smsg.parseMiningNotify(sjob)) {
-      //
-      // mining.notify
-      //
-      convertMiningNotifyStr(line);  // convert mining.notify string
-      sendMiningNotify(line);        // send stratum job to all miners
-     
-      latestJobId_[0]      = latestJobId_[1];
-      latestJobGbtTime_[0] = latestJobGbtTime_[1];
-      latestJobId_[1]      = latestJobId_[2];
-      latestJobGbtTime_[1] = latestJobGbtTime_[2];
-
-      // the jobId always between [0, 9]
-      latestJobId_[2]      = (uint8_t)sjob.jobId_;
-      latestJobGbtTime_[2] = sjob.time_;
-
-      // set last job received time
-      lastJobReceivedTime_ = (uint32_t)time(NULL);
-
-      DLOG(INFO) << "up[" << (int32_t)idx_ << "] stratum job"
-      << ", jobId: "    << sjob.jobId_
-      << ", prevhash: " << sjob.prevHash_
-      << ", version: "  << sjob.version_
-      << ", clean: "    << (sjob.isClean_ ? "true" : "false") << std::endl;
-    }
-    else if (smsg.parseMiningSetDifficulty(&difficulty)) {
-      //
-      // mining.set_difficulty
-      //
-      // just set the default pool diff, than ignore
-      if (poolDefaultDiff_ == 0) {
-        poolDefaultDiff_ = difficulty;
-      }
-    }
-    else if (smsg.parseMiningSetVersionMask(&versionMask)) {
-      //
-      // {"id":null,"method":"mining.set_version_mask","params":["1fffe000"]}
-      //
-      versionMask_ = versionMask;
-    }
-  }
-
-  if (state_ == UP_CONNECTED) {
-    //
-    // {"id":1,"result":[[["mining.set_difficulty","01000002"],
-    //                    ["mining.notify","01000002"]],"01000002",8],"error":null}
-    //
-    uint32_t nonce1 = 0u;
-    int32_t n2size = 0;
-    if (!smsg.getExtraNonce1AndExtraNonce2Size(&nonce1, &n2size)) {
-      LOG(FATAL) << "get extra nonce1 and extra nonce2 failure" << std::endl;
-      return;
-    }
-    extraNonce1_ = nonce1;
-    DLOG(INFO) << "extraNonce1 / SessionID: " << extraNonce1_ << std::endl;
-
-    // check extra nonce2's size, MUST be 8 bytes
-    if (n2size != 8) {
-      LOG(FATAL) << "extra nonce2's size is NOT 8 bytes" << std::endl;
-      return;
-    }
-
-    // subscribe successful
-    state_ = UP_SUBSCRIBED;
-
-    // do mining.authorize and mining.configure (for version rolling)
-    string s = Strings::Format("{\"id\":1,\"method\":\"mining.authorize\",\"params\":[\"%s\",\"\"]}\n"
-                               "{\"id\":2,\"method\":\"mining.configure\",\"params\":[[\"version-rolling\"],"
-                               "{\"version-rolling.mask\":\"ffffffff\",\"version-rolling.min-bit-count\":0}]}\n",
-                               userName_.c_str());
-    sendData(s);
-    return;
-  }
-
-  if (state_ == UP_SUBSCRIBED && smsg.getResultBoolean() == true) {
-    //
-    // check authenticated result
-    // {"error": null, "id": 2, "result": true}
-    //
-    state_ = UP_AUTHENTICATED;  // authorize successful
-    LOG(INFO) << "auth success, name: \"" << userName_
-    << "\", extraNonce1: " << extraNonce1_ << std::endl;
-    return;
-  }
-}
-
 bool UpStratumClient::isAvailable() {
   const uint32_t kJobExpiredTime = 60 * 5;  // seconds
 
   if (state_ == UP_AUTHENTICATED &&
-      latestMiningNotifyStr_.empty() == false &&
       poolDefaultDiff_ != 0 &&
       lastJobReceivedTime_ + kJobExpiredTime > (uint32_t)time(NULL)) {
     return true;
@@ -880,210 +415,6 @@ void StratumSession::recvData(struct evbuffer *buf) {
     handleStratumMessage(line);
   }
 }
-
-void StratumSession::handleStratumMessage(const string &line) {
-  DLOG(INFO) << "recv(" << line.size() << "): " << line << std::endl;
-
-  StratumMessage smsg(line);
-  if (!smsg.isValid()) {
-    LOG(ERROR) << "decode line fail, not a json string" << std::endl;
-    return;
-  }
-
-  string idStr;
-  if (smsg.isStringId()) {
-    idStr = "\"" + smsg.getId() + "\"";
-  } else {
-    idStr = smsg.getId();
-  }
-  if (idStr.empty()) {
-    idStr = "null";
-  }
-
-  if (!smsg.getMethod().empty()) {
-    handleRequest(idStr, smsg);
-    return;
-  }
-
-  // invalid params
-  responseError(idStr, StratumError::ILLEGAL_PARARMS);
-}
-
-void StratumSession::responseError(const string &idStr, int errCode) {
-  //
-  // {"id": 10, "result": null, "error":[21, "Job not found", null]}
-  //
-  char buf[128];
-  int len = snprintf(buf, sizeof(buf),
-                     "{\"id\":%s,\"result\":null,\"error\":[%d,\"%s\",null]}\n",
-                     idStr.empty() ? "null" : idStr.c_str(),
-                     errCode, StratumError::toString(errCode));
-  sendData(buf, len);
-}
-
-void StratumSession::responseTrue(const string &idStr) {
-  const string s = "{\"id\":" + idStr + ",\"result\":true,\"error\":null}\n";
-  sendData(s);
-}
-
-void StratumSession::handleRequest(const string &idStr,
-                                   const StratumMessage &smsg) {
-  const string method = smsg.getMethod();
-  if (method == "mining.submit") {  // most of requests are 'mining.submit'
-    handleRequest_Submit(idStr, smsg);
-  }
-  else if (method == "mining.subscribe") {
-    handleRequest_Subscribe(idStr, smsg);
-  }
-  else if (method == "mining.authorize") {
-    handleRequest_Authorize(idStr, smsg);
-  }
-  else if (method == "mining.configure") {
-    LOG(INFO) << "method: " << method;
-    handleRequest_MiningConfigure(idStr, smsg);
-  } else {
-    // unrecognised method, just ignore it
-    LOG(WARNING) << "unrecognised method: \"" << method << "\"" << std::endl;
-  }
-}
-
-void StratumSession::handleRequest_Subscribe(const string &idStr,
-                                             const StratumMessage &smsg) {
-  if (state_ != DOWN_CONNECTED) {
-    responseError(idStr, StratumError::UNKNOWN);
-    return;
-  }
-  state_ = DOWN_SUBSCRIBED;
-
-  //
-  //  params[0] = client version     [optional]
-  //  params[1] = session id of pool [optional]
-  //
-  // client request eg.:
-  //  {"id": 1, "method": "mining.subscribe", "params": ["bfgminer/4.4.0-32-gac4e9b3", "01ad557d"]}
-  //
-
-  string minerAgent;
-  if (!smsg.parseMiningSubscribe(minerAgent)) {
-    minerAgent = "unknown";
-  }
-
-  // 30 is max length for miner agent
-  minerAgent_ = strdup(minerAgent.substr(0, 30).c_str());
-
-  //
-  // Response:
-  //
-  //  result[0] = 2-tuple with name of subscribed notification and subscription ID.
-  //              Theoretically it may be used for unsubscribing, but obviously miners won't use it.
-  //  result[1] = ExtraNonce1, used for building the coinbase.
-  //  result[2] = Extranonce2_size, the number of bytes that the miner users for its ExtraNonce2 counter
-  //
-  assert(kExtraNonce2Size_ == 4);
-  const uint32_t extraNonce1 = (uint32_t)sessionId_;
-  const string s = Strings::Format("{\"id\":%s,\"result\":[[[\"mining.set_difficulty\",\"%08x\"]"
-                                   ",[\"mining.notify\",\"%08x\"]],\"%08x\",%d],\"error\":null}\n",
-                                   idStr.c_str(), extraNonce1, extraNonce1,
-                                   extraNonce1, kExtraNonce2Size_);
-  sendData(s);
-}
-
-void StratumSession::handleRequest_Authorize(const string &idStr,
-                                             const StratumMessage &smsg) {
-  if (state_ != DOWN_SUBSCRIBED) {
-    responseError(idStr, StratumError::NOT_SUBSCRIBED);
-    return;
-  }
-
-  //
-  //  params[0] = user[.worker]
-  //  params[1] = password
-  //  eg. {"params": ["slush.miner1", "password"], "id": 2, "method": "mining.authorize"}
-  //
-
-  string workerName, fullWorkerName;
-  if (!smsg.parseMiningAuthorize(fullWorkerName)) {
-    responseError(idStr, StratumError::INVALID_USERNAME);
-    return;
-  }
-
-  workerName = getWorkerName(fullWorkerName);  // split by '.'
-  if (workerName.empty())
-    workerName = DEFAULT_WORKER_NAME;
-
-  // auth success
-  responseTrue(idStr);
-  state_ = DOWN_AUTHENTICATED;
-
-  // sent sessionId, minerAgent_, workerName to server_
-  server_->registerWorker(this, minerAgent_, workerName);
-
-  // minerAgent_ will not use anymore
-  if (minerAgent_) {
-    free(minerAgent_);
-    minerAgent_ = NULL;
-  }
-
-  // send mining.set_difficulty
-  server_->sendDefaultMiningDifficulty(this);
-
-  // send latest stratum job
-  server_->sendMiningNotify(this);
-}
-
-void StratumSession::handleRequest_MiningConfigure(const string &idStr,
-                                                   const StratumMessage &smsg) {
-  // mining.configure can be called before mining.authorize and
-  // is usually called before it.
-  /*if (state_ != DOWN_AUTHENTICATED) {
-    responseError(idStr, StratumError::UNAUTHORIZED);
-    return;
-  }*/
-
-  uint32_t versionMask = 0;
-  if (!smsg.parseMiningConfigure(&versionMask)) {
-    responseError(idStr, StratumError::ILLEGAL_PARARMS);
-    return;
-  }
-
-  LOG(INFO) << "handleRequest_MiningConfigure";
-  const uint32_t allowedVersionMask = server_->getVersionMask(upSessionIdx_);
-  versionMask = versionMask & allowedVersionMask;
-
-  string s = Strings::Format("{\"id\":%s,\"result\":{\"version-rolling\":true,\"version-rolling.mask\":\"%08x\"},\"error\":null}\n"
-                             "{\"id\":null,\"method\":\"mining.set_version_mask\",\"params\":[\"%08x\"]}\n",
-                             idStr.c_str(), versionMask, versionMask);
-  sendData(s);
-}
-
-void StratumSession::handleRequest_Submit(const string &idStr,
-                                          const StratumMessage &smsg) {
-  if (state_ != DOWN_AUTHENTICATED) {
-    responseError(idStr, StratumError::UNAUTHORIZED);
-    // there must be something wrong, send reconnect command
-    const string s = "{\"id\":null,\"method\":\"client.reconnect\",\"params\":[]}\n";
-    sendData(s);
-    return;
-  }
-
-  //  params[0] = Worker Name
-  //  params[1] = Job ID
-  //  params[2] = ExtraNonce 2
-  //  params[3] = nTime
-  //  params[4] = nonce
-  Share share;
-  if (!smsg.parseMiningSubmit(share)) {
-    responseError(idStr, StratumError::ILLEGAL_PARARMS);
-    return;
-  }
-
-  // submit share
-  server_->submitShare(share, this);
-
-  responseTrue(idStr);  // we assume shares are valid
-}
-
-
 
 /////////////////////////////////// StratumServer //////////////////////////////
 StratumServer::StratumServer(const string &listenIP, const uint16_t listenPort)
@@ -1139,7 +470,7 @@ void StratumServer::addUpPool(const string &host, const uint16_t port,
   LOG(INFO) << "add pool: " << host << ":" << port << ", username: " << upPoolUserName << std::endl;
 }
 
-UpStratumClient *StratumServer::createUpSession(const int8_t idx) {
+UpStratumClient *StratumServer::createUpSession(int8_t idx) {
   for (size_t i = 0; i < upPoolHost_.size(); i++) {
     struct sockaddr_in sin;
     memset(&sin, 0, sizeof(sin));
@@ -1149,7 +480,7 @@ UpStratumClient *StratumServer::createUpSession(const int8_t idx) {
       continue;
     }
 
-    UpStratumClient *up = new UpStratumClient(idx, base_, upPoolUserName_[i], this);
+    UpStratumClient *up = createUpClient(idx, base_, upPoolUserName_[i], this);
     if (!up->connect(sin)) {
       delete up;
       continue;
@@ -1348,8 +679,11 @@ void StratumServer::listenerCallback(struct evconnlistener *listener,
   uint16_t sessionId = 0u;
   server->sessionIDManager_.allocSessionId(&sessionId);
 
-  StratumSession *conn = new StratumSession(upSessionIdx, sessionId, bev, server,
-                                            ((struct sockaddr_in *)saddr)->sin_addr);
+  StratumSession *conn = server->createDownConnection(upSessionIdx,
+                                                      sessionId,
+                                                      bev,
+                                                      server,
+                                                      reinterpret_cast<struct sockaddr_in *>(saddr)->sin_addr);
   bufferevent_setcb(bev,
                     StratumServer::downReadCallback, NULL,
                     StratumServer::downEventCallback, (void*)conn);
@@ -1500,13 +834,13 @@ void StratumServer::upEventCallback(struct bufferevent *bev,
   server->removeUpConnection(up);
 }
 
-void StratumServer::sendMiningNotifyToAll(const int8_t idx, const string &notify) {
+void StratumServer::sendDataToAll(const int8_t idx, const string &data) {
   for (size_t i = 0; i < downSessions_.size(); i++) {
     StratumSession *s = downSessions_[i];
     if (s == NULL || s->upSessionIdx_ != idx)
       continue;
 
-    s->sendData(notify);
+    s->sendData(data);
   }
 }
 
@@ -1515,7 +849,7 @@ void StratumServer::sendMiningNotify(StratumSession *downSession) {
   if (up == NULL || up->latestMiningNotifyStr_.length() == 0)
     return;
 
-  downSession->sendData(up->latestMiningNotifyStr_);
+  downSession->sendMiningNotify(up->latestMiningNotifyStr_);
 }
 
 void StratumServer::sendDefaultMiningDifficulty(StratumSession *downSession) {
@@ -1523,10 +857,7 @@ void StratumServer::sendDefaultMiningDifficulty(StratumSession *downSession) {
   if (up == NULL)
     return;
 
-  const string s = Strings::Format("{\"id\":null,\"method\":\"mining.set_difficulty\""
-                                   ",\"params\":[%" PRIu32"]}\n",
-                                   up->poolDefaultDiff_);
-  downSession->sendData(s);
+  downSession->sendMiningDifficulty(up->poolDefaultDiff_);
 }
 
 void StratumServer::sendMiningDifficulty(UpStratumClient *upconn,
@@ -1535,9 +866,7 @@ void StratumServer::sendMiningDifficulty(UpStratumClient *upconn,
   if (downSession == NULL)
     return;
 
-  const string s = Strings::Format("{\"id\":null,\"method\":\"mining.set_difficulty\""
-                                   ",\"params\":[%" PRIu64"]}\n", diff);
-  downSession->sendData(s);
+  downSession->sendMiningDifficulty(diff);
 }
 
 int8_t StratumServer::findUpSessionIdx() {
@@ -1558,76 +887,6 @@ int8_t StratumServer::findUpSessionIdx() {
     }
   }
   return idx;
-}
-
-void StratumServer::submitShare(const Share &share,
-                                StratumSession *downSession) {
-  UpStratumClient *up = upSessions_[downSession->upSessionIdx_];
-
-  bool hasVersionMask = share.versionMask_ != 0;
-  bool isTimeChanged = true;
-  if ((share.jobId_ == up->latestJobId_[2] && share.time_ == up->latestJobGbtTime_[2]) ||
-      (share.jobId_ == up->latestJobId_[1] && share.time_ == up->latestJobGbtTime_[1]) ||
-      (share.jobId_ == up->latestJobId_[0] && share.time_ == up->latestJobGbtTime_[0])) {
-    isTimeChanged = false;
-  }
-
-  //
-  // | magic_number(1) | cmd(1) | len (2) | jobId (uint8_t) | session_id (uint16_t) |
-  // | extra_nonce2 (uint32_t) | nNonce (uint32_t) | [nTime (uint32_t)] | [nVersionMask (uint32_t)] |
-  //
-  uint16_t len = 15;
-  if (isTimeChanged) len += 4;
-  if (hasVersionMask) len += 4;
-
-  uint8_t cmd = (isTimeChanged && hasVersionMask) ? CMD_SUBMIT_SHARE_WITH_TIME_VER :
-                (isTimeChanged)  ? CMD_SUBMIT_SHARE_WITH_TIME :
-                (hasVersionMask) ? CMD_SUBMIT_SHARE_WITH_VER :
-                /* default */      CMD_SUBMIT_SHARE;
-
-  string buf;
-  buf.resize(len, 0);
-  uint8_t *p = (uint8_t *)buf.data();
-
-  // cmd
-  *p++ = CMD_MAGIC_NUMBER;
-  *p++ = cmd;
-
-  // len
-  *(uint16_t *)p = len;
-  p += 2;
-
-  // jobId
-  *p++ = (uint8_t)share.jobId_;
-
-  // session Id
-  *(uint16_t *)p = downSession->sessionId_;
-  p += 2;
-
-  // extra_nonce2
-  *(uint32_t *)p = share.extraNonce2_;
-  p += 4;
-
-  // nonce
-  *(uint32_t *)p = share.nonce_;
-  p += 4;
-
-  // ntime
-  if (isTimeChanged) {
-    *(uint32_t *)p = share.time_;
-    p += 4;
-  }
-
-  // nVersionMask
-  if (hasVersionMask) {
-    *(uint32_t *)p = share.versionMask_;
-    p += 4;
-  }
-
-  assert(p - (uint8_t *)buf.data() == (int64_t)buf.size());
-
-  // send buf
-  up->sendData(buf);
 }
 
 void StratumServer::registerWorker(StratumSession *downSession,
@@ -1702,8 +961,4 @@ void StratumServer::unRegisterWorker(StratumSession *downSession) {
 
   UpStratumClient *up = upSessions_[downSession->upSessionIdx_];
   up->sendData(buf);
-}
-
-uint32_t StratumServer::getVersionMask(uint8_t upSessionIdx) {
-  return upSessions_[upSessionIdx]->getVersionMask();
 }
