+++ src/agent/agent_conf.json
@@ -1,4 +1,5 @@
 {
+    "agent_type": "btc",
     "agent_listen_ip": "0.0.0.0",
     "agent_listen_port": 3333,
     "pools": [
